// src/app/core/interfaces/deployStatus.ts
export interface DeployStatus {
  crq:              string;
  nomb_estado:      string;
  url:              number | string;
  l_user:           string;
  estadoinput:      null | string;
  estado:           string;
  fecha_liberacion: Date;
  reliberado:       string;
  reversado:        string;
}